+++ pdf-chatbot/app.py
@@ -2,87 +2,142 @@ import argparse
 import os
 import shutil
 from time import sleep
-from typing import Any, List
+from typing import Dict, Any, List, Optional
 
+import faiss
 import gradio as gr
-from langchain.chains import ConversationalRetrievalChain
-from langchain.memory import ConversationBufferMemory
-from langchain.text_splitter import CharacterTextSplitter
-from langchain_core.output_parsers import StrOutputParser
-from langchain_core.prompts import ChatPromptTemplate
-from langchain_core.output_parsers import StrOutputParser
-from langchain_core.runnables import RunnablePassthrough
-from langchain_community.embeddings import HuggingFaceBgeEmbeddings
-from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler
-from langchain_community.llms import HuggingFaceEndpoint
-from langchain_community.llms import VLLM
-from langchain_community.vectorstores import FAISS
-from PyPDF2 import PdfReader
+
+from llama_index.embeddings.huggingface import HuggingFaceEmbedding
+from llama_index.vector_stores.faiss import FaissVectorStore
+
+from llama_index.core import QueryBundle
+from llama_index.core.node_parser import SentenceSplitter
+from llama_index.core.retrievers import BaseRetriever
+from llama_index.core.schema import TextNode, NodeWithScore
+from llama_index.core.vector_stores import VectorStoreQuery
+from llama_index.core.vector_stores.types import VectorStoreQueryMode
+from llama_index.core.chat_engine import ContextChatEngine
+from llama_index.llms.vllm import Vllm
+from llama_index.llms.huggingface import HuggingFaceLLM
+from llama_index.readers.file import PyMuPDFReader
+
 import torch
 
-def get_pdf_text(pdf_docs):
-    """
-    Extract text from a list of PDF documents.
+CHAT_TEMPLATE = """
+{%- set ns = namespace(first_system=false) -%}
+{{-'<s>'-}}
+{%- for message in messages %}
+    {%- if message['role'] == 'system' -%}
+        {{-' [INST] ' + message['content'] + '\n\n'-}}
+        {%- set ns.first_system = true -%}
+    {%- else -%}
+        {%- if message['role'] == 'user' -%}
+            {%- if ns.first_system -%}
+                {{-'' + message['content'].rstrip() + ' [/INST] '-}}
+                {%- set ns.first_system = false -%}
+            {%- else -%}
+                {{-' [INST] ' + message['content'].rstrip() + ' [/INST] '-}}
+            {%- endif -%}
+        {%- else -%}
+            {{-'' + message['content'] + '</s>' -}}
+        {%- endif -%}
+    {%- endif -%}
+{%- endfor -%}
+{%- if add_generation_prompt -%}
+    {{-''-}}
+{%- endif -%}
+"""
 
-    Parameters
-    ----------
-    pdf_docs : list
-        List of PDF documents to extract text from.
+def generate_vector_store_nodes(pdf_doc_path: str, embed_model: HuggingFaceEmbedding):
+    loader = PyMuPDFReader()
+    documents = loader.load(file_path=pdf_doc_path)
+    text_parser = SentenceSplitter(chunk_size=1536, chunk_overlap=384)
 
-    Returns
-    -------
-    str
-        Extracted text from all the PDF documents.
+    text_chunks = []
+    # maintain relationship with source doc index, to help inject doc metadata later
+    doc_indices = []
+    for doc_idx, doc in enumerate(documents):
+        cur_text_chunks = text_parser.split_text(doc.text)
+        text_chunks.extend(cur_text_chunks)
+        doc_indices.extend([doc_idx] * len(cur_text_chunks))
 
-    """
-    text = ""
-    for pdf in pdf_docs:
-        pdf_reader = PdfReader(pdf)
-        for page in pdf_reader.pages:
-            text += page.extract_text()
-    return text
+    nodes = []
+    for idx, text_chunk in enumerate(text_chunks):
+        node = TextNode(
+            text=text_chunk,
+        )
+        src_doc = documents[doc_indices[idx]]
+        node.metadata = src_doc.metadata
+        nodes.append(node)
 
+    for node in nodes:
+        node_embedding = embed_model.get_text_embedding(
+            node.get_content(metadata_mode="all")
+        )
+        node.embedding = node_embedding
 
-def get_text_chunks(text):
-    """
-    Split the input text into chunks.
+    return nodes
 
-    Parameters
-    ----------
-    text : str
-        The input text to be split.
+class FaissVectorDBRetriever(BaseRetriever):
+    """Retriever over a postgres vector store."""
 
-    Returns
-    -------
-    list
-        List of text chunks.
+    def __init__(
+        self,
+        vector_store: FaissVectorStore,
+        embed_model: HuggingFaceEmbedding,
+        query_mode: str = "default",
+        similarity_top_k: int = 2,
+    ) -> None:
+        self._vector_store = vector_store
+        self._embed_model = embed_model
+        self._query_mode = query_mode
+        self._similarity_top_k = similarity_top_k
+        super().__init__()
 
-    """
-    text_splitter = CharacterTextSplitter(
-        separator="\n", chunk_size=1500, chunk_overlap=300, length_function=len
-    )
-    chunks = text_splitter.split_text(text)
-    return chunks
+    def _retrieve(self, query_bundle: QueryBundle) -> List[NodeWithScore]:
+        query_embedding = self._embed_model.get_query_embedding(
+            query_bundle.query_str
+        )
+        vector_store_query = VectorStoreQuery(
+            query_embedding=query_embedding,
+            similarity_top_k=self._similarity_top_k,
+            mode=VectorStoreQueryMode(self._query_mode),
+        )
+        query_result = self._vector_store.query(vector_store_query)
+        if query_result.nodes is None:
+            return []
+
+        nodes_with_scores = []
+        for index, node in enumerate(query_result.nodes):
+            score: Optional[float] = None
+            if query_result.similarities is not None:
+                score = query_result.similarities[index]
+            nodes_with_scores.append(NodeWithScore(node=node, score=score))
+
+        return nodes_with_scores
 
 class RAGInterface:
     def __init__(
         self,
         embedding_model_name: str,
-        encode_kwargs: dict,
-        vllm_kwargs: dict,
         docs_folder: str = "./docs",
+        stream: bool = False,
+        use_flash_attention: bool = False,
+        use_vllm: bool = True,
+        vllm_kwargs: Optional[Dict[str, Any]] = {},
     ):
         self.device = "cuda" if torch.cuda.is_available() else "cpu"
-        self.embedding_model = HuggingFaceBgeEmbeddings(
-            model_name=embedding_model_name, encode_kwargs=encode_kwargs, model_kwargs={"device": self.device}
-        )
-        self.encode_kwargs = encode_kwargs
-        self.vectorstore: FAISS = None
+        self.embedding = HuggingFaceEmbedding(model_name=embedding_model_name, device=self.device)
+        self.faiss_index = faiss.IndexFlatL2(1024) # 1024 is dimension of the embeddings
+        self.vector_store = FaissVectorStore(faiss_index=self.faiss_index)
         self.docs_folder = docs_folder
-        self.vllm_kwargs = vllm_kwargs
+        self.stream = stream
+        self.use_flash_attention = use_flash_attention
+        self.use_vllm = use_vllm
+        self.vllm_kwargs = vllm_kwargs if use_vllm else {}
         print(f"Using accelerator: {self.device}")
 
-    def initialize_conversation_chain(self, initial_docs: List[str], model_name: str = "mistralai/Mistral-7B-Instruct-v0.2"):
+    def initialize_chat_engine(self, initial_docs: List[str], model_name: str = "mistralai/Mistral-7B-Instruct-v0.2"):
         """
         Initialize a vector database from a list of PDF documents.
 
@@ -93,50 +148,44 @@ class RAGInterface:
 
         """
 
-        print(f"Scanning all pdf files in {len(initial_docs)} Documents...")
+        print(f"Initializing vector database from {len(initial_docs)} Documents...")
+        for pdf_file_path in initial_docs:
+            nodes = generate_vector_store_nodes(pdf_file_path, self.embedding)
+            self.vector_store.add_nodes(nodes)
 
-        raw_text = get_pdf_text(initial_docs)
-        if raw_text == "":
-            raw_text = "Initial text"
-        text_chunks = get_text_chunks(raw_text)
+        if self.use_vllm:
+            # Note: vLLM does not support streaming interface yet
+            # ref: https://github.com/run-llama/llama_index/issues/9477
+            print(f"Loading LLM from {model_name} using vLLM...")
+            llm = Vllm(
+                model=model_name,
+                trust_remote_code=True,  # mandatory for hf models
+                max_new_tokens=4096,
+                vllm_kwargs=self.vllm_kwargs if self.vllm_kwargs else {},
+                top_k=10,
+                top_p=0.95,
+                temperature=0.8,
+            )
+        else:
+            print(f"Loading LLM from {model_name} using transformers.AutoModelForCausalLM...")
+            model_kwargs = {"temperature": 0.8, "do_sample": True, "top_k": 10, "top_p": 0.95}
+            if self.use_flash_attention:
+                model_kwargs["attn_implementation"] = "flash_attention_2"
+            llm = HuggingFaceLLM(
+                model_name=model_name,
+                tokenizer_name=model_name,
+                max_new_tokens=4096,
+                is_chat_model=True,
+                model_kwargs=model_kwargs,
+            )
+            # Overwrite chat_template to support system prompt
+            llm._tokenizer.chat_template = CHAT_TEMPLATE
 
-        print("Initializing vector database...")
-        self.vectorstore = FAISS.from_texts(texts=text_chunks, embedding=self.embedding_model)
-
-        print("Initializing conversation chain...")
-        llm = HuggingFaceEndpoint(
-            model=model_name,
-            streaming=True,
-            repetition_penalty=1.2,
-            do_sample=True,
-        )
-
-        memory = ConversationBufferMemory(memory_key="chat_history", return_messages=True)
-
-        self.conversation = ConversationalRetrievalChain.from_llm(
-            llm=llm, retriever=self.vectorstore.as_retriever(), memory=memory
-        )
-
-        # template = """Answer the question based on the following context and reference.
-        # If there is no context given, answer the question directly without saying there is no context.
-
-        # Context: {context}
-
-        # Question: {question}
-        # """
-        # prompt = ChatPromptTemplate.from_template(template)
-        # self.conversation = (
-        #     {
-        #         "context": self.vectorstore.as_retriever(),
-        #         "question": RunnablePassthrough(),
-        #     }
-        #     | prompt
-        #     | llm
-        #     | StrOutputParser()
-        # )
+        self.retriever = FaissVectorDBRetriever(self.vector_store, self.embedding, query_mode="default", similarity_top_k=2)
+        self.chat_engine = ContextChatEngine.from_defaults(retriever=self.retriever, llm=llm)
 
     def add_document(self, list_file_obj: List, progress=gr.Progress()):
-        if self.vectorstore is None:
+        if self.vector_store is None:
             raise ValueError("Vectorstore not initialized. Please run initialize_database() first.")
 
         if list_file_obj is None:
@@ -147,26 +196,29 @@ class RAGInterface:
         pdf_docs = [x.name for x in list_file_obj if x is not None]
         for pdf in pdf_docs:
             shutil.copy(pdf, self.docs_folder)
-        gr.Info("Extracting text from PDFs...")
-        raw_text = get_pdf_text(pdf_docs)
-        gr.Info("Splitting text into chunks...")
-        text_chunks = get_text_chunks(raw_text)
-        gr.Info("Adding chunks to vector database...")
-        self.vectorstore.add_texts(texts=text_chunks)
+
+        gr.Info("Adding documents into vector database...")
+        for pdf_file_path in pdf_docs:
+            nodes = generate_vector_store_nodes(pdf_file_path, self.embedding)
+            self.vector_store.add_nodes(nodes)
+            progress(1, desc=f"Adding {pdf_file_path} to vector database")
 
         gr.Info("Upload Completed!")
         return gr.update(value="Upload PDF documents", interactive=True)
 
-    def get_retriever(self):
-        if self.vectorstore is None:
-            raise ValueError("Vectorstore not initialized. Please run initialize_database() first.")
-        return self.vectorstore.as_retriever()
-
     def handle_chat(self, message, history):
-        full_reponse = ""
-        for response in self.conversation.stream(message):
-            full_reponse += response
-            yield full_reponse
+        if self.stream:
+            streaming_response = self.chat_engine.stream_chat(message)
+            full_response = ""
+            for token in streaming_response.response_gen:
+                full_response += token
+                yield full_response
+
+            return full_response
+        else:
+            chat_response = self.chat_engine.chat(message)
+            return chat_response.response
+
 
 def close_app():
     gr.Info("Terminated the app!")
@@ -191,15 +243,21 @@ def main(args: argparse.Namespace):
 
     ragger = RAGInterface(
         embedding_model_name=args.embedding_model_name,
-        encode_kwargs={"normalize_embeddings": True},
-        vllm_kwargs={"max_model_len": int(args.vllm_max_model_len), "enforce_eager": args.vllm_enforce_eager},
-    )
-    ragger.initialize_conversation_chain(initial_docs, model_name=args.model_name)
+        stream=True if args.stream else False,
+        use_flash_attention=True if args.use_flash_attention else False,
+        use_vllm=True if args.use_vllm else False,
+        vllm_kwargs={
+            "max_model_len": int(args.vllm_max_model_len),
+            "enforce_eager": args.vllm_enforce_eager,
+        } if args.use_vllm else {},
 
-    with gr.Blocks(css=css, title="PDF Chatbot with LangChain🦜 and Open-source LLMs") as demo:
+    )
+    ragger.initialize_chat_engine(initial_docs, model_name=args.model_name)
+
+    with gr.Blocks(css=css, title="PDF Chatbot with LlamaIndex🦙 and Open-source LLMs") as demo:
         with gr.Row():
             gr.Markdown(
-            f"""<h2>PDF Chatbot with LangChain🦜 and {args.model_name}</h2>
+            f"""<h2>PDF Chatbot with LlamaIndex🦙 and {args.model_name}</h2>
             <h3>Ask any questions about your PDF documents, along with follow-ups</h3>
             <b>Note:</b> This AI assistant performs retrieval-augmented generation from your PDF documents.<br>
             Initial documents are loaded from the `{args.docs_folder}` folder. You can add more documents by clicking the button below.<br>
@@ -236,10 +294,21 @@ if __name__ == "__main__":
     parser.add_argument("--docs-folder", default="./docs")
     parser.add_argument("--embedding-model-name", default="BAAI/bge-m3")
     parser.add_argument("--model-name", default="TheBloke/Mistral-7B-Instruct-v0.2-AWQ")
+    parser.add_argument("--use-vllm", action="store_true")
+    parser.add_argument("--stream", action="store_true")
+    parser.add_argument("--use-flash-attention", action="store_true")
     parser.add_argument("--vllm-max-model-len", default=4096)
     parser.add_argument("--vllm-enforce-eager", action="store_true")
     parser.add_argument("--hf-token", default="")
 
     args = parser.parse_args()
+
+    # Validate arguments
+    if args.use_vllm:
+        print("[WARN] Using vLLM for generation. You might want to install vLLM with `pip install -U vllm llama-index-llms-vllm` before running the app.")
+        if args.stream:
+            raise ValueError("vLLM on LlamaIndex does not streaming interface yet. Please remove --stream flag.")
+    if args.use_flash_attention:
+        print("[WARN] Using Flash Attention for vLLM. You might want to install Flash Attention with `FLASH_ATTENTION_FORCE_BUILD=TRUE pip install -U flash-attn` before running the app.")
 
     main(args)
