--- pdf-chatbot/requirements-llamaindex.txt
@@ -1,11 +1,0 @@
-autoawq==0.2.3
-faiss-cpu==1.7.4
-flash-attn==2.5.6
-gradio==4.15.0
-llama-index==0.10.19
-llama-index-embeddings-huggingface==0.1.4
-llama-index-llms-huggingface==0.1.4
-llama-index-llms-vllm==0.1.6
-llama-index-vector-stores-faiss==0.1.2
-vllm==0.3.3
-transformers==4.38.0
